Removed another dependency on castracker.
We had previously been using a casetracker function for populating the timetracker user select list. Now casetracker is optional and if not present will pull all users from the Drupal database. In the future, it might be nice to have a view that controls what users are available for time entry attribution.